requirements.txt
aioredis==1.2.0
async-timeout==3.0.1
attrs==18.2.0
boto3==1.9.50
botocore==1.12.50
certifi==2018.1.18
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
colorama==0.4.1
configparser==3.5.0
constantly==15.1.0
cryptography==2.7
cycler==0.10.0
daphne==2.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2.2
django-allauth==0.36.0
django-ckeditor==5.6.1
django-classy-tags==0.9.0
django-cors-headers==2.3.0
django-decouple==2.1
django-filer==1.5.0
django-formtools==2.1
django-heroku==0.3.1
django-js-asset==1.1.0
django-mptt==0.10.0
django-polymorphic==2.0.3
django-rest-auth==0.9.3
django-sekizai==1.0.0
django-storages==1.7.1
django-treebeard==4.3
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.5
djangocms-admin-style==1.2.9
djangocms-attributes-field==1.1.0
djangocms-column==1.9.0
djangocms-file==2.3.0
djangocms-googlemap==1.3.0
djangocms-installer==1.1.0
djangocms-link==2.4.0
djangocms-picture==2.0.8
djangocms-snippet==2.1.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.7.0
djangocms-video==2.0.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
gunicorn==19.9.0
hiredis==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
inflection==0.3.1
isort==4.3.20
itsdangerous==1.1.0
jmespath==0.9.3
joblib==0.13.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
matplotlib==3.1.0
mccabe==0.6.1
more-itertools==5.0.0
msgpack==0.6.1
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
ndg-httpsclient==0.5.1
numpy==1.16.4
oauthlib==2.1.0
pandas==0.24.2
paramiko==2.6.0
Pillow==6.0.0
pipenv==2018.11.26
psycopg2==2.8.3
pyasn1==0.4.5
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.6.4
pylint==2.3.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.0
PySocks==1.7.0
python-dateutil==2.7.5
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
Quandl==3.4.8
requests==2.22.0
requests-oauthlib==1.0.0
s3transfer==0.1.13
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
selenium==3.141.0
six==1.11.0
sklearn==0.0
sqlparse==0.3.0
storage==0.0.4.3
stripe==2.29.0
typed-ast==1.4.0
tzlocal==1.5.1
Unidecode==1.0.23
urllib3==1.22
virtualenv==16.7.5
virtualenv-clone==0.5.3
webencodings==0.5.1
whitenoise==4.1.3

